fix usage of JID.user
