+++ lib/session.js
@@ -4,7 +4,7 @@ var util = require('util')
   , tls = require('tls')
   , EventEmitter = require('events').EventEmitter
   , Connection = require('node-xmpp-core').Connection
-  , JID = require('node-xmpp-core').JID
+  , JID = require('node-xmpp-jid')
   , SRV = require('node-xmpp-core').SRV
   , BOSHConnection = require('./bosh')
   , WSConnection = require('./websockets')
