+++ package.json
@@ -26,12 +26,13 @@
     "url": "https://github.com/node-xmpp/node-xmpp-client/issues"
   },
   "dependencies": {
-    "request": "~2.55.0",
-    "faye-websocket": "~0.7.0",
-    "node-xmpp-core": "^3.0.0",
     "browser-request": "~0.3.1",
+    "debug": "~1.0.0",
+    "faye-websocket": "~0.7.0",
     "minimist": "0.0.8",
-    "debug": "~1.0.0"
+    "node-xmpp-core": "^3.0.0",
+    "node-xmpp-jid": "^1.0.0",
+    "request": "~2.55.0"
   },
   "devDependencies": {
     "brfs": "0.0.8",
@@ -47,7 +48,7 @@
     "grunt-mocha-cli": "~1.3.0",
     "grunt-mocha-istanbul": "^2.1.0",
     "grunt-mocha-phantomjs": "^2.0.0",
-    "node-xmpp-server": "^1.0.2",
+    "node-xmpp-server": "^2.0.0",
     "should": "~2.1.1"
   },
   "browser": {
